+++ POC_Client/Pages/ScreenPlacementSelectingPage.xaml.cs
@@ -53,8 +53,8 @@ public partial class ScreenPlacementSelectingPage : ContentPage
     protected override async void OnSizeAllocated(double i_Width, double i_Height)
     {
         base.OnSizeAllocated(i_Width,i_Height);
-        m_pageLogic.SetPlayerScreenSize(i_Width, i_Height);
-        
+        m_pageLogic.SetPlayerScreenSize((int)i_Width, (int)i_Height);
+       
     }
 
     private void initializeButtons()
